Refs #47: Convert to SCSS files for proper import
The prior fix switched to SCSS methods for importing and broke
how the CSS files are imported. This switches to SCSS methods
across the board to unify with how this is done throughout Foreman.

Additionally, the original fix did not entirely fix the problem.
The sat6.css file included a mangled copy of the angular-ui-select
minimized CSS that was breaking the SCSS builder form being able
to process it. This change updates that code to match the upstream
minimized copy of the CSS for that version of angular-ui-select.